refactor(utils): move egomimicUtils symbols into the existing utils modules
egomimicUtils.py was a 1,120-line grab-bag. Rather than inventing new modules,
the 20 symbols that still have callers move into the modules that already
exist and already own that concern:

  pose_utils.py    +9  pose/transform/camera-frame + the two array interpolators
  viz_utils.py     +3  draw_actions, draw_dot_on_frame, get_gaze_endpoint
  tensor_utils.py  +2  get_sinusoid_encoding_table, EinOpsRearrange
  utils.py         +4  str2bool, fmt, download_from_huggingface, STD_SCALE
  metrics.py       +2  frechet_gaussian_over_time, reverse_kl_from_samples (NEW --
                       the only case with no existing home; both are eval metrics
                       consumed solely by eval/eval_hpt.py)

viz_utils already imported draw_actions/draw_dot_on_frame/get_gaze_endpoint from
egomimicUtils, so moving them there removes an import rather than adding one.

Code is moved verbatim (exact AST source spans). egomimicUtils.py re-exports the
moved symbols, so all existing importers keep working unchanged.

Imports are resolved, not copied: each destination gets exactly the third-party
imports its moved code references, cross-module references point at the real new
home, and any destination that previously imported a now-local symbol from
egomimicUtils has that import removed -- otherwise egomimicUtils re-exporting from
viz_utils while viz_utils imports from egomimicUtils is a circular import.

The 24 symbols with no caller anywhere are left in egomimicUtils untouched;
removing them is a separate decision and a separate commit.

Verified: ruff F821/F811 clean on egomimic/utils; all 20 moved symbols resolve
through the old path; zero unresolvable global references in any moved function
body (checked by walking each bodys AST against its new module globals -- the
check that a plain import test does not catch); 13/14 importer modules import,
the 14th being egomimic.robot.rollout which fails on a pre-existing unrelated
bare robot_utils import.
